refactor: create router package, and don't define interface in controller
