+++ cmd/sample-api/main.go
@@ -4,12 +4,13 @@ import (
 	"net/http"
 
 	"github.com/podhmo-sandbox/sample-api/controller"
+	"github.com/podhmo-sandbox/sample-api/controller/router"
 	"github.com/podhmo-sandbox/sample-api/model/repository"
 )
 
 var tr = repository.NewTodoRepository()
 var tc = controller.NewTodoController(tr)
-var ro = controller.NewRouter(tc)
+var ro = router.NewRouter(tc)
 
 func main() {
 	server := http.Server{
